Add short delay between DMing votes.
A problem, especially in the last AGM, is that the bot has been ratelimited while sending DMs to everyone.
I believe this delay (0.5s) will fix this, as the ratelimit resets fairly often. I don't know exactly how long of a delay is needed.

Turnout is around 25/45 for E/AGM respectively, so this will only add 10-30s per role.